+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `AppController::isBanned()` method renamed as `isSpammer()`;
 * updated for `me-tools` 2.18.7;
 * added `MeCms\ORM\PostAndPageEntity` abstract class;
+* added `MeCms\TestSuite\BannersAndPhotosAdminControllerTestCase` abstract class;
 * `PostsAndPagesTables` class moved to `MeCms\ORM` namespace;
 * updated for `php-tools` 1.2.5.